gestione ecdcezioni

+++ Lezione_18/src/main/java/com/epicode/spring/service/UtenteService.java
@@ -10,6 +10,9 @@ import org.springframework.stereotype.Service;
 import com.epicode.spring.model.Utente;
 import com.epicode.spring.repository.UtenteDAORepository;
 
+import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
+
 @Service
 public class UtenteService {
 	
@@ -30,20 +33,35 @@ public class UtenteService {
 	}
 	
 	public Utente getUtente(Long id) {
+		if(!uRepo.existsById(id)) {
+			throw new EntityNotFoundException("L'utente riuchiesto è inesistente!");
+		}
 		return uRepo.findById(id).get();	
 	}
 	
 	public Utente creaUtenteCustom(Utente utente) {
+		if(uRepo.existsByCognome(utente.getCognome())) {
+			throw new EntityExistsException("Questo cognome esiste già!");
+		}
 		return uRepo.save(utente);
 	}
 	
 	public String eliminaUtente(Long id) {
+		if(!uRepo.existsById(id)) {
+			throw new EntityNotFoundException("L'utente riuchiesto è inesistente!");
+		}
 		Utente u = uRepo.findById(id).get();
 		uRepo.delete(u);
 		return "Utente eliminato";
 	}
 	
-	public Utente aggiornaUtente(Utente u) {
+	public Utente aggiornaUtente(Utente u, Long id) {
+		if(!uRepo.existsById(id)) {
+			throw new EntityNotFoundException("L'utente riuchiesto è inesistente!");
+		}
+		if(id != u.getId()) {
+			throw new EntityNotFoundException("L'id e l'id dell'utente non corrispondono!");
+		}
 		return uRepo.save(u);
 	}
